+++ build.sh
@@ -10,3 +10,7 @@ cmake -DCMAKE_BUILD_TYPE=Release -DRTMIDI_API_JACK=OFF -DRTMIDI_BUILD_TESTING=OF
 # cmake -DCMAKE_BUILD_TYPE=Release -DRTMIDI_BUILD_TESTING=OFF ..
 
 make
+
+# Remove unnecessary files after build done
+# rm -r externals/ 
+# rm -rf CMakeLists.txt build.sh line.cpp